Serializing tests running while still fast.
Tests are consciously run one at a time (not all at parallel and
relying on single-thread model to run only one at a time), while
pauses are still minimal because of more workers thrown
into setTimeout queue.

This is preparing for running async tests (which need to be
properly serialized).