Well-known variables must be exempt of 'globally undefined'.
This stops recompilation in CLI. Moreover, we have two
platforms: jQuery and window are known in one, but
process and global are known in the second. That means some methods
would always be uncompilable (but question is in which platform).
This is not normal.

There should be a pragma for "this is known variable".
I would propose `<var knownVar>`.