Wrap Javascript objects before showing
Nicolas reminded me that JS objects can't respond to __asString(), and
so must be wrapped. This is done via _st(), which will wrap the object
in a JSObjectProxy. Initial tests seem to indicate that this is enough -
I was able to print the Window object and some other JS objects.