Add some convenience messages to HashedCollection
HashedCollection already has a message for instance creation from a
collection of associations: #fromPairs. As @herby mentioned in #433,
this message is poorly named, as the associations are already "paired"
up.

I've renamed this message to #newFrom, and it accepts a collection of
associations. I've also added a message for creating a HashedCollection
or one of its derivatives from a flattened set of associations called
\#newFromPairs. The message accepts an array where every 2 elements form
an association - the odd element being the key and the even the value.
Of course this means that #newFromPairs only accepts arrays that have an
even size.