change locating the bower_components directory
The bower_components directory is usually found
in the root directory of the application.
However, it is not necessarily found in the Amber
directory.

Usecase:

1. create new project
2. run 'bower init'
3. run 'bower install --save amber#master' // use master for the requirejs stuff
4. create index.html and load Amber via
       <script type='text/javascript' src='bower_components/amber/support/amber.js'></script>
    <script type='text/javascript' src='bower_components/amber/support/requirejs/require.min.js'></script>

5. load index.html via a file:/// url

Without this change all bower components will get
searched in file:///project_dir/bower_components/amber/bower_components/.
With this change dependent files are loaded correctly.