BUGFIX: Inspecting a string in Helios breaks into the debugger.
Evaluating `'hello' inspect` in a workspace would break into
the debugger. The problematic code seems to be in `HLListWidget>>refresh`
which seeks to make visible the currently selected item, without
checking whether there is a such an item. Adding an `ifNotNil:`
condition before that action fixes the problem.

Confession: I'm not aware of when exactly the `#ensureVisible:` part
gets triggered.