Some minor speedups plus StringSet class.
In Helios, Ctrl+space, f, r (second use after previous one) was hell slow.
The cause was Set of all selectors.
Swapping it for StringSet gave maybe 100x sppedup (I don't see
the thing slowing me a bit - before it was a 2-3 seconds).

Hopefully it solves another big UI bug of Helios - when writing
a selector, I immediately write space after comma, but it was ignored
because it again looked over selector to give me some list to choose
or whatnot. It was _big_ PITA.