`Smalltalk` is global variable, not a class
This is in par with Pharo, where `Smalltalk` is
(single) instance of `SmalltalkImage`, and system things
are accessed via `Smalltalk xxx`.

This also open possibility to create `Smalltalk vm`
to access low-level options and `Smalltalk globals`
as a (JSObjectProxy-ed) dictionary of globals,
thus making Amber API more compatible with other Smalltalks.