Add a Browser service
It's nice to be able to browse an object's class by doing `aThing browse`
in a workspace. Creating a method that opens the Helios browser is very
easy, but then Object relies generally on Helios, and specifically on the
current implementation of Helios.

A Browser service has been created, and anything that considers itself a
browser can register. Just make sure your browser responds to `open` and
`openOn:`. Objects can be browsed by telling them to `browse`.

The Helios Browser and older IDE Browser will register themselves, with
the IDE Browser deferring to Helios if it exists.