Find references to a thing if browsing fails
The Browser should try and `DoIt` a line or selection before browsing the
returned object's class. If an error occurs, then the line probably wasn't
valid Smalltalk, and may be a text snippet, so a References browser is
opened.

Behavior is undefined if your code contains an underscore (_) that isn't
in a string or comment.