Cleaner exporting code for class names in javascript.
Fixed the nil subclasses which are now outputted as subclasses of JS `null`,
not of `globals.nil`, which was undefined (so it worked as a side effect).

Also, code clearly states when it wants ST name and when JS name
(so there is no `'globals.', stClassName` in exporter methods any more).